Fixed Properties use for Ship Components and fixed a bug where the mouse couldn't be used until the gamemode was changed for the first time (bounds were not set)

+++ se.liu.ida.l2.tddd78.psc/src/weaponry/projectile/MissileProjectile.java
@@ -9,7 +9,7 @@ public class MissileProjectile extends AbstractProjectile
 {
 
 	public MissileProjectile(final float selfX, final float selfY, final float targetX, final float targetY,
-							 final float velocity, final Starship enemyShip, final int damageOnImpact, final int blastRadius)
+							 final float velocity, final Starship enemyShip, final float damageOnImpact, final int blastRadius)
 	{
 		super(selfX, selfY, targetX, targetY, velocity, enemyShip, damageOnImpact, blastRadius);
 	}
